fix: サーバーでしか使用しない関数をserverファイルに移動
