+++ app/.server/loaders/rankingLoader/index.ts
@@ -1,6 +1,6 @@
 import { type LoaderFunctionArgs, json } from '@remix-run/cloudflare'
 import { getTodayDate } from '~/lib/date.server'
-import { loadEnvironment } from '~/lib/utils'
+import { loadEnvironment } from '~/lib/utils.server'
 import { getAvatarRanking, getClothRanking } from '~/module/get/get-ranking'
 import { createClient } from '~/module/supabase/create-client.server'
