+++ app/routes/sign-out.tsx
@@ -1,5 +1,5 @@
 import { type ActionFunctionArgs, redirect } from '@remix-run/cloudflare'
-import { loadEnvironment } from '~/lib/utils'
+import { loadEnvironment } from '~/lib/utils.server'
 import { createClient } from '~/module/supabase/create-client.server'
 
 export const action = async ({ context, request }: ActionFunctionArgs) => {
